backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
PyJWT==2.8.0
bcrypt==4.1.3
python-dotenv==1.0.1
pydantic[email]==2.7.1
pydantic-settings==2.2.1
scikit-learn==1.4.2
numpy==1.26.4
joblib==1.4.2
python-multipart==0.0.9
slowapi==0.1.9
alembic==1.13.1
